+++ devito/dse/backends/basic.py
@@ -44,7 +44,7 @@ class BasicRewriter(AbstractRewriter):
         skip = [e for e in cluster.exprs if e.lhs.base.function.is_Array]
         candidates = [e for e in cluster.exprs if e not in skip]
 
-        make = lambda: Scalar(name=template()).indexify()
+        make = lambda: Scalar(name=template(), dtype=cluster.dtype).indexify()
 
         processed = common_subexprs_elimination(candidates, make)
